Change TableDrawer and TableTreeDrawer
