--- Packages/UGF.RuntimeTools/Editor/Tables/TableTreeDrawerEditorUtility.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 205f1944680243b4b623837edfa55d3f
-timeCreated: 1704719431
